fix conflicts in pg_proc
I removed the prodataaccess labeling of some built-in functions. We don't
use prodataacess for anything ATM, but if we did, we'd need to put them
back in. Right now, however, I think we're better off just sticking as
close to the upstream code as possible. If we wanted to use the field for
anything, we'd need to label a whole bunch of other built-in functions
anyway.