All the conflicts in pg_depend.c were because of CaQL.
On further investigation, it turns out that *all* the differences to
upstream in that file were because of CaQL, so I just took the whole
file from upstream as is.

Well, almost: there was also one instance of changing an old
heap_formtuple() call to the new heap_form_tuple() interface. I re-did
that change. But other than that, the file is now 100% identical to
upstream.