Fix conflicts in src/backend/utils/misc
I also decided to remove checkpoint.h, and move the definitions in it
to bgwriter.h. Having a "checkpoint.h", corresponding to "checkpoint.c",
makes perfect sense, but those function definitions are in bgwriter.h in
PostgreSQL, and keeping the code as close to upstream as possible trumps
the consistency of keeping definitions for "foo.c" in header file called
"foo.h".

This also makes it possible to enable autovacuum again. I added a comment
on why it's still disabled by default, even though the default was changed
to 'on' in PostgreSQL 8.3.