fix conflicts in smgr.
CheckPointStartLock was replaced in upstream with the PGPROC->inCommit flag.
The CHECKPOINT_START_LOCK macros in cdbpersistentstore.h need to be adjusted
accordingly, but that's not totally straightforward as we have to ensure
that inCommit is somehow reset on error, while the lwlock was released
automatically. So I didn't do that right now. Instead, I removed the
CHECKPOINT_START_LOCK/UNLOCK calls from smgrDoDeleteActions, because they
seemed redundant: According to the comment, their purpose was to make sure
we acquire the locks in right order, but we never block on setting the
inCommit flag, so we no longer need to worry about that. I just hope that
PersistentRelation_Dropped is indeed acquiring the lock/flag correctly,
as the comment said.