Fix grammar merge conflicts, and resolve grammar conflicts.
NOTE: This removes the support for "CLUSTER VERBOSE ..." syntax. The
VERBOSE option was always ignored, so this is no loss in functionality,
you just can't add that noiseword there anymore. According to git history,
this was added to the syntax back in 2009, for "PostgreSQL compatibility",
but AFAICS PostgreSQL has never had VERBOSE option for CLUSTER, so I have
no idea what that was all about. It's possibly worth noting in release
notes as a backwards-incompatibility, though.