Remove references to HEAP_HASEXTENDED.
The HEAP_COMPRESSED flag was removed from heap tuple header. ISTM that
all of these checks for it are not actually needed. And it's probably only
a good thing to send already-compressed datums in compressed format, to
reduce network traffic, rather than decompress them in the sender before
sending. There was even a comment implying that. I'm not sure why it was
not done that way to begin with, though. There may well be some code
elsewhere that doesn't expect to get compressed Datums from the
deserialization routine. We'll find out eventually...