Remove unnecessary 'need_free' argument from defGetString().
All of the callers are in places where leaking a few bytes of memory to
the current memory context will do no harm. Either parsing, or processing
a DDL command, or planning. So let's simplify the callers by removing the
argument. That makes the code match the upstream again, which makes merging
easier.

These changes were originally made to reduce the memory consumption when
doing parse analysis on a heavily partitioned table, but the previous
commit provided a more whole-sale solution for that, so we don't need to
nickel-and-dime every allocation anymore.

Conflicts:
	src/backend/commands/typecmds.c
	src/backend/parser/analyze.c