Fix conflicts in autovacuum.c
I resolved almost all of them by taking the upstream version. I think
we had backported much of this already, but there were a lot of whitespace
and other cosmetic differences that caused the conflicts. Also, some
parts were already backported from a later version, as the SHMEM_OFFSET
stuff was missing. I took the 8.3 version of that too, essentially reverting
the earlier backport, but there is no functional difference so that should
be fine.