Fix counting of tupletable slots for subplans, by getting rid of the counting.
Backport this upstream patch, which replaces the tuple table from an
array to a linked list. With that, we don't need to know the size of the
array beforehand, and don't need to count the slots.

There was something wrong with the way the slots were counted in GDPB,
probably caused by the merge of the upstream changes that changed the way
SubPlan nodes work. All subplans are now initialized as a separate step,
rather than as part of the main plan tree. Instead of trying to track down
the bug there, let's just get rid of the need for counting the slots in
the first place.

commit f92e8a4b5ee6a22252cbba012d629f5cefef913f
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Sun Sep 27 20:09:58 2009 +0000

    Replace the array-style TupleTable data structure with a simple List of
    TupleTableSlot nodes.  This eliminates the need to count in advance
    how many Slots will be needed, which seems more than worth the small
    increase in the amount of palloc traffic during executor startup.

    The ExecCountSlots infrastructure is now all dead code, but I'll remove it
    in a separate commit for clarity.

    Per a comment from Robert Haas.