Fix assertion failure at VACUUM FULL during initdb.
I tried to rearrange vacuum_rel() and friends so that the transaction
boundaries would be more like in upstream, but that turned out to be
difficult. So as a quick fix, just teach Begin- and
EndNonTransactionalInvalidation to cope with the different pattern.
The old-style VACUUM FULL will go away in later versions of PostgreSQL,
so I'm not inclined to spend too much time cleaning it up now. Then again,
refactoring vacuum_rel() back to closer to upstream would be nice anyway.
Maybe some day..