Further fixes to XID handling.
Move the GetTopTransactionIdIfAny() call after the triggers have been fired.
A trigger might cause an XID to be assigned.

Clear TopTransactionStateData.transactionId, so we don't try to clear
MyProc->xid twice, tripping an assertion in ProcArrayEndtransaction(), if
something goes wrong in the late stages of AbortTransaction(), and we recurse
back to AbortTransaction().