Move memoryAccount field from PlannedStmt to QueryDesc.
PlannedStmt is supposed to be immutable, and is kept across executions
and even across transactions, for cached plans. There are more
GPDB-added fields in there that are modified, but this is a start.

Fixes a crash from re-executing a cached plan, caused by memoryAccount
field pointing to a now-invalid object after the first execution.