Move slice table, transientRecordTypes, and IntoClause's oidInfo out of PlannedStmt.
These are transient information, created in ExecutorStart, not something
that should be cached along with the plan. transientRecordTypes and oidInfo
in particular were stored in PlannedStmt only so that they can be conveniently
dispatched to QEs along with the plan. PlannedStmt ought to be immutable,
or you get strange errors when the same PlannedStmt is cached and reused
across executions. So create a separate QueryDispatchDesc struct to hold
information that's not supposed to be cached with a plan, and is dispatched.

Fixes remaining crashes from plancache regression test.