Make TableFunctionScan more like FunctionScan, and don't fetch the RTE.
There was some confusion over the RTE entry, and in subselect.c, the
RTE was of type RTE_SUBQUERY, in one of the regression tests. I'm not
sure why that happened, some sort of rearrangement of the range table
after the TableFunctionScan node was constructed perhaps. But let's make
this more like regular FunctionScan node anyway, and store the few
fields we need from the RTE in the TableFunctionScan node directly.