Make the InitPlan Param mechanism cope with InitPlans that are never executed.
If there are two InitPlan nodes inside each other, and the outer InitPlan
never executes the inner InitPlan (because the value of the outer InitPlan
was determined without it), don't get confused when some of the params
are not set (= won't even have a valid type OID).

bfv_subquery regression tests exercised this.