Fix "plan should not reference subplan's variable" errors.
With nested subqueries, the SS_finalize_plan() call for the inner subquery
was happening too early, when make_subqueryscan called it. It's only needed
for some GPDB-specific callers, so move the SS_finalize_plan() to the
callers that need it, so that it's not called in wrong order in other
cases.