Partial backport of b153c09 from PostgreSQL 8.4
This is a partial backport of a larger body of work which also already have
been partially backported. Rather than taking all of the upstream code this
is filling in some blanks to get proper line number error messages.

The parts left out are mainly the changes to the node functions where the
current GPDB code is kept (for now).

  commit b153c0920960a6059b67969469166fb29c0105d7
  Author: Tom Lane <tgl@sss.pgh.pa.us>
  Date:   Mon Sep 1 20:42:46 2008 +0000

    Add a bunch of new error location reports to parse-analysis error messages.
    There are still some weak spots around JOIN USING and relation alias lists,
    but most errors reported within backend/parser/ now have locations.