Further distributed transaction management fixes.
I think I understand a bit more about this again. The QE reader's
top-level XID needs to be set, so that TransactionIdIsCurrentTransaction()
considers the XID as "current". Otherwise a QE reader will not see
tables/rows created earlier in the same transaction, as evidenced by
the "relation not found" errors from rangefuncs regression test. However,
the QE reader needs special care at commit time, as we don't want it to
write a commit record. And more acutely, it shouldn't fail the assertion
in ProcArrayEndTransaction(), when MyProc->xid is not set, even though
TopTransactionStateData.transactionId is.

A simple test case for was:

begin;
create table foo (id int);
insert into foo values (1), (2);
commit;

Somewhat surprisingly, you didn't get an error with just one inserted row,
as there's a fastpath for that in the executor that happened to avoid the
problem. But with a two-row insert, you got an error.