Rework error context printing for sortby operators
This adds error context to "operator << is not a valid ordering operator"
error messages. Moving to passing the SortBy struct to addTargetToSortList()
moves us closer to upstream where that was done in a later version. One
perhaps controversial change is that it now prints the error context on the
SELECT rather than the sort field, neither of which is technically correct
since it should indicate on the sort operator in question. This however use
the SortBy.location field which is more correct so fixing it to properly
indicate on the operator should be quite possible with this structure.