Fix the order that commands are executed when creating a partitioned table.
The order the NOTICEs come out (see gpdist.out change) is still different
from what it used to be, because the statements are now transformed in
depth-first order. But we add them to the list of commands to execute
in the same order as before. This fixes the error about missing parent
partition, when creating a table with subpartitions.