Add support for hashing enums in table distribution
This adds support for enum types to be hashed based on their Oid
similar typ regtype Oid which may or may not be an ideal solution.
It seems to work though so starting with this.

Also fixes the enum test failures by removing the RI checks. Since
Greenplum doesn't enforce RI and we have tests for that elsewhere
it seems a bit superflous to run even more tests on it.