Don't transform already-transformed ALTER TABLE statements in segments.
The master sends AlterTableStmts in already-transformed form, so don't
try to transform them again. The concrete problem from this was a
"relation "rank" is not a partitioned" error from partition1 regression
test. Apparently we don't have all the pg_partition_rule rows in the
segments, so the segment could not do all the same validations as the
master did. I'm a bit surprised by that, but transforming a statement
again certainly doesn't seem right, and that makes the error go away,
so I guess we're good.