Put back code that was left out of the merge of upstream b17b7fae.
This was causing "ALTER TABLE ... ADD COLUMN ... DEFAULT NULL" to
do an unnecesary table rewrite (it still causes a rewrite on appendonly-
tables, that's intentional, but not on heap tables). Also caused 'rowtypes'
regression test to fail, because it did an ALTER that only worked as long
as no rewrite was done.