Remove the limitation that a default must be provided for ADD COLUMN on AO table
We had later added code that allows "DEFAULT NULL", by doing a table
rewrite. DEFAULT NULL is really the same as no default, so we might as well
do a table rewrite for that case too, and save the code needed to handle
them differently.

The reason to do this now is that the default_is_null flag wasn't being
set correctly. Might as well just get rid of it, instead of debugging why.