Don't create array types for the auxiliary heap created for a bitmap index.
The immediate problem with it was that the array type's OID was not
dispatched from the master to the segments, so it was assigned different
OIDs in segments. And that caused OID conflicts with concurrent activity,
as happened with the regression tests (duplicate key violation on
"pg_type_oid_index").

We could've sent the array type's OID from master to segments, like we do
for regular tables, but these things don't really need array types, so
seems better to not create them at all.