Fix portals-related regression tests.
There are a few limitation in GPDB: backwards scan are not supported,
you cannot update the distribution key, and the order of rows without
order by is not well defined (it's not on Postgres either, but doesn't
change in practice). And doing "UPDATE .. WHERE OF c" twice in a row
doesn't work; the second UPDATE does nothing. But these are all
pre-existing issues.