Cleanup of 'percentile' regression test.
Adjust the expected output for changes in error context lines. Also reformat
some of the test queries, and make sure the formatting is the same between
the .sql file and .out file (pg_regress uses "diff -w", which masks out some
of the differences, but it was annoying while doing a manual diff).