Fixs some "fewer pathkeys than join clauses" errors.
Some expressions might not be present in the new target list, and we should
just use the old expression in that case. I'm actually not 100% sure when
that happens, but that's what the functiond did before Equivalence Classes
merge too, and with this fix, things seem to work.

This isn't a complete fix, there are still some failures like this in the
bfv_planner test case.