Fix "bogus varno 65001" errors from EXPLAIN.
The way we dig the displayed name for a Var was changed in upstream.
get_variable now digs into the plan tree, until it finds the original
Scan node where the column came from. That mechanims didn't work for
ShareInputScan nodes, where we had removed the subplan. The subplan
is removed from all but one of the ShareInputScan nodes, to turn the
plan tree from a DAG into a tree, as is expected by set_plan_references()
and possibly other functions.

To fix, mangle the targetlist of ShareInput nodes, using fake CTE
rangetable entries that act as placeholders for the subplan's targetlist
entries. The table/column names you get in the EXPLAIN output are not
as nice with this method as they used to be, but it's good enough.