Fix remaining "fewer pathkeys than join clauses" errors.
If the GROUP BY clause contains expressions that are in fact unique,
and don't need to be sorted (like in "WHERE c = 123 GROUP BY c"), it's
normal that parse->groupClause contains more items than the current
sorting pathkey. We should be looking at root->group_pathkeys instead,
which has a one-to-match correspondence with root->groupClause.