Fix remove_unused_initplans().
I had just disabled it earlier, because it was causing crashes, but
we "no parameter found for initplan subquery" errors were caused by
the lack of it. The problem is that if the initplan's Params are not
removed, we get confused at execution time, when we try to dispatch
the parameter's value to the executor nodes, and the Param hasn't
been set. So we still need remove_unused_initplans().

The problem with remove_unused_initplans() was that it didn't know
about the global list of subplans. Even if we remove all references of
an initplan from elsewhere, we still try to initialize the subplan
at executor startup. Motion nodes get upset if we try to call
ExecInitNode() on them, but it hasn't been assigned a slice ID. And
we only assign slice IDs to Motion nodes that are referenced from
the main plan. To fix, replace unused initplans in the global
list with a dummy false Result node, which is happy to be initialized
without any special preprocessing.