Fix occasional PANIC when replaying heap insert records for pg_type.
Every operation that creates a WAL record on one of the "persistent tables"
needs to have this information. Without it, the XLOG_HEAP2_CLEAN record was
not replayed in phase 3 of startup, and a subsequent insert failed because
there was not enough free space on the page.

We really should have an assertion or something to catch bugs of omission
like this...