Fix GetSnapshotData(), mangled by the merge.
Some of the old code should've been replaced with new version, but we
instead had both versions. Also update comments to match the new
upstream version.

The practical change here is that we no longer assign an XID unnecessarily
to a transaction, just because it acquires a snapshot. (This came up while
trying to fix CREATE INDEX CONCURRENTLY, as some parts of that assert
that the transaction doesn't have an XID assigned to it.)