Mark bitmap index AM as not capable of "col IS NULL" searches.
Marking that as true in the merge was wishful thinking. It probably
wouldn't take a lot of effort to make it actually support that. It
already indexes NULLs (or at least it's marked with amindexnulls='t' in
pg_am), so it just needs to be taught about ISNULL scankeys. See
commit c8076f09 that did the same for B-tree in the upstream. But now
is not the time for any new feature development.