Fix predicate propagation with outer joins.
Predicate propagation would incorrectly derive a top-level clause from a
clause that came from the "OUTER JOIN ON (..)" clause. That's not legal.
join_gp test case was failing because of that. I'm not entirely sure
why this stopped working now, rather than earlier when the Equivalence
Classes patch was merged. There were some outer-join related changes
in this batch, though.

Fix by keeping track of the "ojscope" for each RestrictInfo, and
propagating it to any derived RestrictInfos. Also change the way we
find all the "relevant" clauses to derive new quals from. Instead of
scanning all rels for join- and base-restrictinfos after they've been
distributed, memorize any clauses that not true equivalence clauses to
a separate list when we do the normal qual distribution. Clauses that
are handled by the equivalence class machinery are not interesting as
base for predicate propagation, because the EC machinery can already
derive all useful quals from those.