When operating on a partitioned table, only keep the parent table locked.
When creating the partitions in CREATE TABLE, or the indexes on the
partitions in CREATE INDEX, or when dropping a partitioned table, release
the locks on the partitions early. We rely on the lock on the parent
table to prevent other backends from modifying the partitions.

This reduces the number of locks held at the same time, if a table has
a lot of partitions. You might reach max_locks_per_transactions limit
otherwise.

This was broken during the merge.

In the passing, clean up the partition_locking test case, which tests for
this. At one point while I hacked on this, I had a version that released
the locks early in the master, but held them 'til the end of transaction
in the segments. The 'locktest' view as it was wouldn't have revealed
that, so make it also show all the locks in segment 0.