I changed the rule for naming the index backing a PK earlier. Recant.
PostgreSQL always uses the constraint's name for the index that backs
the constraint. We had changed that in GPDB, so that the index name is
always independently of the constraint name, as if no name was given
for the constraint. I reverted that change earlier as part of the merge,
so that the behaviour for top-level tables was the same as upstream, but
partitions kept the old GPDB behaviour (the indexes in different partitions
must have different names, because you can't have two indexes with the same
name). But I have to recant on that. It made ALTER TABLE SPLIT PARTITION to
fail, because the constraints of different partitions had different names.
We perhaps could've salvaged it partially, and made sure that the constraint
has the same name in all partitions, but named the indexes differently, but
it seems better to just go back to the old scheme altogether.

I don't remember what problem I was trying to fix by changing the behaviour
earlier, but I'm not seeing any new problems attributable to this now.

This patch makes one extra behavioural change: If you do CREATE TABLE foo
(LIKE bar WITH CONSTRAINTS), the copied constraints in the new table will
have the same names as the originals. That was not possible in the upstream,
because there the constraint's name must match the index's name, so you'd
get a conflict, but because we don't have that connection in GPDB, we can
do it. I'm marketing it as a feature here, but it really just fell out of
the implementation automatically. It's easier to always copy the name than
to try to figure out if we're creating a partition (in which case we must
do it).