Fix "incompatible loci in target inheritance set" errors.
cdb_set_cheapest_dedup() function turned the "dummy join" Append with
zero subplans plan into a more complicated one. We don't want that, because
is_dummy_join() will not recognize the more complicated plan as dummy, which
caused the check in inheritance_planner() to fail. Not sure why this started
to happen now, but the fix seems straightforward.