Backport plan cache invalidation improvements from PostgreSQL 8.4.
PostgreSQL 8.3 included plan cache invalidation support, but only for
relations. GPDB contained a hack to invalidate all plans in master between
transactions, for the same purpose, but the proper plan cache invalidation
is a much better solution. However, because the GPDB hack also handled
dropped/recreated functions in some cases, if we just leave it out and
replace with the 8.3 plan cache invalidation, there will be a regression
in those cases. To fix, backport the rest of the plan cache invalidation
support from 8.4, which handles functions and operators too.

The upstream commit was:

commit ee33b95d9c2ecec170bc517783d7268a4bd0c793
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Tue Sep 9 18:58:09 2008 +0000

    Improve the plan cache invalidation mechanism to make it invalidate plans
    when user-defined functions used in a plan are modified.  Also invalidate
    plans when schemas, operators, or operator classes are modified; but for these
    cases we just invalidate everything rather than tracking exact dependencies,
    since these types of objects seldom change in a production database.

    Tom Lane; loosely based on a patch by Martin Pihlak.