Fix conflicts in contrib/xml2
This didn't compile at all prior to the merge so I guess it's not
used but since the autoconf options are live and the code there it
should at the very least not break the build. Further testing will
need to be done to verify that it indeed does anything useful.

The tuplestore in xpath_table() is created with random access even
though it might not require it, the scaffolding for dealing with
this (commit 05bba3d176e0adc1a032d5c8c6ea2a7622c7dd0d) came in 8.4
and is yet to be merged. For now do the safe but potentially less
performant option.