Fix partition_locking test case to not be sensitive on # of segments.
I changed the test earlier to print the locks on segments too, but to
make the test output not dependent on the # of segments, I only printed
the locks held in segment 0. But that's not good enough: if a single
row is inserted to a table, whether or not the lock is held in segment 0
depends on whether the row happened to land on that segment, which depends
on the hash function and number of segments in the cluster. Fix by
printing a summary of locks across all segments.