Perform meatball surgery on ORCA to make it compile
The translation code from Planned Statement to DXL is slated for
removal which should be done as part of the 8.3 merge to minimize
the footprint. This is clearly not the best way to go about it
since there are dependencies on the code still but for now lets
see if we can get the PULSE builds further until the QO team has
time to help us out.