Re-plan prepared statements on every EXECUTE, like we used to.
In GPDB 4.3, we used to re-plan the query with the given parameters every
time a prepared statement is EXECUTEd. In PostgreSQL, the query is planned
at PREPARE, and the same plan is reused on EXECUTE, and we accidentally
lost the GPDB changes in this area in the merge and fell back to the
PostgreSQL behaviour. Refactor the code in ExecuteQuery(), by adding a new
RevalidateCachedPlanWithParams() function, which allows re-planning a
cached plan with specific parameters. Normally, it wouldn't make much sense
to cache a plan using the plancache functions, and then not use the cached
plan; you might as well just call pg_parse_analyze() and pg_plan_query()
directly. But this allows keeping the code as close to upstream as possible.

This also refactors the handling of CREATE TABLE AS with EXECUTE, to keep
that code as close to upstream as possible as well.

This fixes the failure in direct_dispatch test case. A prepared statement
was not direct-dispatched because the planner could not determine at
PREPARE time which segment the target row belongs to, but at EXECUTE time,
with the parameter's value, it could.