Avoid deadlock on catchup interrupt.
An earlier attempt at this checked AmIInSIGUSR1Handler() to see if we
are currently processing a catchup event. But that's not good enough:
we also process catchup interrupts outside the signal handler, in
EnableCatchupInterrupt(). I saw lockups during "make installcheck-good"
with a stack trace that shows a backend waiting for lock on a temporary
relation, while trying to truncate it when committing the transaction
opened for processing a catchup event.

For reference, the commit message for the commit that introduced the
AmIInSIGUSR1Handler check said:

    Recent parallel installcheck-good revealed we have a chance to process
    catchup interrupt while waiting for commit-prepare, and if the prepared
    transaction has created a temporary table with on commit option, the
    newly opened transaction for the sake of AcceptInvalidationMessages()
    cannot see and fails before the commit-prepare.  It's even not clear if
    we are safe to open and commit another transaction between prepare and
    commit-prepare, but for now just skip the oncommit operation as it
    doesn't have any effect anyway.