Don't call ExecAssignScanProjectionInfo while in partitionMemoryContext.
I saw a segfault in the buildfarm, from the ExecAssignScanProjectionInfo()
call later in the function, from cleanup_slot(). I think what happened
is that the old slot was allocated in the first
ExecAssignScanProjectionInfo() call, in partitionMemoryContext, and
then immediately destroyed when the memory context was reset. The second
call to ExecAssignScanProjectionInfo() tries to free the slot, again,
causing the segfault.

I'm not quite sure why that hasn't been a problem before, though. The code
hasn't changed in the merge. In any case, calling
ExecAssignScanProjectionInfo() twice doesn't seem useful.

In the passing, clean up the code a bit. I found having separate variables,
indexState and scanState, which point to the same struct, to be confusing.