Fix out-of-bounds write to scanTupleSlot
ss_ScanTupleSlot is not an array, it's a single slot. This used to be mostly
harmless before the merge, as the slots were allocated in one large array,
so in practice ss_ScanTupleSlot always pointed to an element in a the large
array. Accessing ss_ScanTupleSlot[1] was bogus then too, because that slot
might not have anything to do with the DynamicTableScan node, but after
the merge, all slots are is simply palloc'd, so doing ss_ScanTupleSlot[1]
will clobber whatever happens to be at that memory location.

I found this when debugging a segfault from DynamicTableScan node. I'm not
100% if this was the cause of that segfault, but certainly seems wrong.