Remove duplicated bitgetpage() function.
The one in nodeBitmapHeapScan.c is what the upstream has. The one in
execBitmapHeapScan.c was copied from it in GPDB, but that was not kept
to up-to-date with the merge, and didn't know anything about HOT, for
example.

It's not clear to me why we have the execBitmapHeapScan.c file at all,
why not just use the functions in nodeBitmapHeapScan.c. But I'll leave
investigating that for another day.