fix conflict in fd.h
In GPDB, we had changed the File* functions to use int64 as the type for
file offset. In upstream, it was now changed from "long" to "off_t".

Could we assume that off_t is 64-bits wide these days? Not sure. For now,
I kept the GPDB version, with int64.