Fix conflicts in hashfunc.c and hash AM.
NOTE: this brings in the contents of hashfunc.c from the head of
REL8_4_STABLE. That includes later changes to the hash algorithm used. I
didn't feel like trying to fix it to an intermediate state that we would
just rewrite again later.

The hash function had been replaced in GPDB, too, but I couldn't quite
figure out what the GPDB algorithm was, and whether it was better or how.
In any case, I believe the new PostgreSQL algorithm is decent, so let's
just use that. I'm not very impressed by the old code, there was weird
stuff going on with the little and big endianess stuff. And at the top,
WORDS_BIGENDIAN was misspelled as WORS_BIGENDIAN, so it never worked as
intended on big endian systems.

Note that GPDB uses a completely different set of hash functions for
calculating the DISTRIBUTED BY key, so this doesn't affect pg_upgrade.
This does invalidate hash indexes, but they're not supported on GPDB
anyway. And we don't support hash partitioning either.