Fix conflicts in tuplesort and tuplestore.
NOTE: tuplesort_mk.c is a bit confused with trace_sort and dtrace. It
had dtrace probes inside "if (trace_sort)" checks, but didn't normally
log anything even with trace_sort. I kept the probes, but made them
unconditional, so that they're enabled even without the trace_sort GUC.
To follow the example of tuplesort.c, it would be nice to add some actual
logging there with trace_sort, that doesn't require dtrace. But this will
do for now.